Agent, Island: remove/rename system info collection infrastructure
System info collectors got replaced with credential collectors. Infrastructure in the code needs to be renamed accordingly
