+++ monkey/common/common_consts/credential_collector_names.py
