--- monkey/infection_monkey/system_info/collectors/__init__.py
@@ -1,3 +1,0 @@
-"""
-This package holds all the dynamic (plugin) collectors
-"""
