--- monkey/infection_monkey/telemetry/system_info_telem.py
@@ -1,20 +1,0 @@
-from common.common_consts.telem_categories import TelemCategoryEnum
-from infection_monkey.telemetry.base_telem import BaseTelem
-
-
-class SystemInfoTelem(BaseTelem):
-    def __init__(self, system_info):
-        """
-        Default system info telemetry constructor
-        :param system_info: System info returned from SystemInfoCollector.get_info()
-        """
-        super(SystemInfoTelem, self).__init__()
-        self.system_info = system_info
-
-    telem_category = TelemCategoryEnum.SYSTEM_INFO
-
-    def get_data(self):
-        return self.system_info
-
-    def send(self, log_data=False):
-        super(SystemInfoTelem, self).send(log_data)
